Fix error when creating 0 instances
Why would anyone want to create zero instances?

Suppose there are templates being re-used for multiple environments, and in
a development or test environment, multiple major application roles/functions
(let's call them "foo" and "bar") are handled by a single instance to save on
costs. In production, however, each major application role lives on its own
instance(s) for better resiliency.

Since Terraform doesn't support conditional invocation of modules,
specifying an `instance_count` of `0` for "foo" and "bar" in dev/test
and an `instance_count` of `1` for "foobar", while specifying an
`instance_count` of `1` for "foo" and "bar", respectively, and `0` for
"foobar", would permit reuse across environments and avoid a lot of
copy-paste that may drift.

**Note:** The SSM document association portion was the only part of this
template that was currently producing errors, so it's the only one I
addressed. This will technically still create all of the other base
resources like roles, iam policies, etc. We can tie those to
`instance_count` as well, if desired.